% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmdreg-package.R
\docType{package}
\name{cmdreg-package}
\alias{cmdreg}
\alias{cmdreg-package}
\title{cmdreg: Managing regression environment}
\description{
TBA.
}
\author{
\strong{Maintainer}: Hirotake Ito \email{hohoge@gmail.com}

}
\keyword{internal}
