% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression_analysis_environment.R
\docType{data}
\name{RegressionAnalysisEnvironment}
\alias{RegressionAnalysisEnvironment}
\title{R6 Class representing a Regression analysis environment}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
RegressionAnalysisEnvironment
}
\arguments{
\item{...}{arguments becoming object attribute.}

\item{name}{attributuion name.}

\item{value}{attributuion value}

\item{error_stop}{logical. If error_stop is TRUE, this function will stop when any error happens.}

\item{...}{arguments becoming variable of `analyze_base` function.}
}
\value{
character string representing this object.

A new `RegressionAnalysisEnvironment` object.
}
\description{
Return the object name.

Create a new RegressionAnalysisEnvironment object.

attributuion setter.

this function is base function for analysis. You should override this function at the child class.

this function is base function for analysis. You should override this function at the child class.
}
\details{
A environment have `analyze` method.
}
\section{Fields}{

\describe{
\item{\code{status}}{represents status of analysis.}

\item{\code{messeage}}{represents messeage of analysis.}

\item{\code{result}}{is object representing result of analysis.}

\item{\code{document}}{is Document.}

\item{\code{tag}}{is the list of tag representing this object.}
}}

\keyword{datasets}
